Bug fixing

+++ src/components/JobList.js
@@ -4,6 +4,8 @@ import SalesforceRestApi from "@ocdla/salesforce/SalesforceRestApi.js";
 import Job from "@ocdla/employment/Job.js";
 import Component from "./Component.js";
 import JobCard from "./JobCard.js";
+import Intro from "./Intro.js";
+
 
 
 
@@ -36,42 +38,23 @@ export default class JobList extends Component {
 
   async onRequestDelete(dataset) {
     let id = dataset.id;
-    let resp = await this.api.delete("Job__c", id);
+    let resp;
+    if (window.confirm("Are you sure you want to delete this?")) {
+      resp = await this.api.delete("Job__c", id);
+    } else {
+      return false;
+    }
 
-    if(resp.ok) {
+    if(resp === true) {
       const e = new CustomEvent("rerender", { detail: this });
       document.dispatchEvent(e);
+      return true;
     }
     else
     {
-      window.alert("An error occurred while deleting the record.");
+      throw new Error("An error occurred while deleting the record.");
     }
-
   }
-
-
-  // async onRequestDelete(job) {
-  //   let message;
-  //   let userId = USER_ID;
-
-  //   if (!job.isOwner(userId)) {
-  //     throw new Error("You don't have permission to perform this action.");
-  //   }
-
-  //   try {
-  //     await this.deleteJob(job.id);
-  //     message = "The record was deleted.";
-  //   } catch (e) {
-  //     message = e.message;
-  //   }
-
-  //   await this.loadData(this.records);
-
-  //   return;
-  // }
-
- 
-
 
   async loadData() {
       let resp = await this.api.query(QUERY);
@@ -84,11 +67,16 @@ export default class JobList extends Component {
     //let message = props.error || props.message || "";
     let message = "";
     let fullWidth = this.fullWidth;
+    let intro = Intro();
 
     return (
       <div {
         ...fullWidth ? ({class: "full-container"}) : ("")
       }>
+        <div class="intro hiddenIntro">
+          {intro}
+        </div>
+
         <div style="color:red;" class="error">
           {message}
         </div>
